first and last commit Css styles
Added the body and its a text shadow and is for the left and right borders.Added the background image that have an id of tabs and that is in the navigation that is in the header and in the body. Added the image to be centered both horizontally and vertically. The rule is for the navigations that are on the left side of the website and they are vertical navigations. The rule creates a border and the border radius to make the edges round. Added the image on the webpage that is center and the rule is giving a radial gradient, border and rounded edges. Added the block-quotes and the rule is giving the quotes an background image and has round repeat and the size is 100 percent. ready for review